chore: Rename app
