+++ google_pathways_web_app/app.py
@@ -1,10 +1,10 @@
 from flask import Flask
 from flask_migrate import Migrate
 
-from google_pathways_api.config.config import ConfigurationFactory
-from google_pathways_api.db.models import db
-from google_pathways_api.resources.pathways import pathways_blueprint
-from google_pathways_api.resources.static_resources import static_blueprint
+from google_pathways_web_app.config.config import ConfigurationFactory
+from google_pathways_web_app.db.models import db
+from google_pathways_web_app.resources.pathways import pathways_blueprint
+from google_pathways_web_app.resources.static_resources import static_blueprint
 
 
 def create_app():
