+++ meta.json
@@ -5,7 +5,7 @@
     "support": {
         "email": "",
         "github": "",
-        "documentation": "https://github.com/brighthive/google-pathways-api"
+        "documentation": "https://github.com/brighthive/google-pathways-web-app"
     },
     "ops": {
         "email": "",
